Add Travis CI to automatically run flake8 tests
Step 1 of #818

There are tons of tools in the [GitHub Marketplace](https://github.com/marketplace) but the one that I use the most is [Travis CI](https://travis-ci.org) because it is free for Open Source projects like this one and it is quite flexible at running tests.  This config file has Travis run [flake8](http://flake8.pycqa.org) tests on every pull request so that you can quickly see if there are breaking changes.  At https://travis-ci.org/profile you would need to flip the switch for this repo on and the commit any change to the repo for the testing to begin.